refactor(ThumbnailSizeContext): Add custom hook to access context
